+++ jwst/associations/tests/test_level3_product_names.py
@@ -99,19 +99,6 @@ def test_level3_productname_components_acid():
     assert matches['opt_elem'] == 'f560w'
 
 
-def test_level35_names(pool_file):
-    rules = registry_level3_only()
-    pool = AssociationPool.read(pool_file)
-    asns = generate(pool, rules)
-    for asn in asns:
-        product_name = asn['products'][0]['name']
-        if asn['asn_rule'] == 'Asn_IFU':
-            m = re.match(LEVEL3_PRODUCT_NAME_NO_OPTELEM_REGEX, product_name)
-        else:
-            m = re.match(LEVEL3_PRODUCT_NAME_REGEX, product_name)
-        assert m is not None
-
-
 def test_level3_names(pool_file, global_constraints):
     rules = registry_level3_only(
         global_constraints=global_constraints